Add Kafka base properties hierarchy
