+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/kafka/KafkaProperties.java
@@ -16,32 +16,19 @@
 
 package org.springframework.boot.autoconfigure.kafka;
 
-import java.io.IOException;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
-import java.util.Locale;
 import java.util.Map;
-
-import org.apache.kafka.clients.CommonClientConfigs;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.config.SslConfigs;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.kafka.common.serialization.StringSerializer;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.DeprecatedConfigurationProperty;
 import org.springframework.boot.context.properties.PropertyMapper;
-import org.springframework.boot.context.properties.source.MutuallyExclusiveConfigurationPropertiesException;
 import org.springframework.boot.convert.DurationUnit;
-import org.springframework.core.io.Resource;
 import org.springframework.kafka.listener.ContainerProperties.AckMode;
 import org.springframework.kafka.security.jaas.KafkaJaasLoginModuleInitializer;
-import org.springframework.util.CollectionUtils;
 import org.springframework.util.unit.DataSize;
 
 /**
@@ -55,31 +42,11 @@ import org.springframework.util.unit.DataSize;
  * @author Artem Bilan
  * @author Nakul Mishra
  * @author Tomaz Fernandes
+ * @author Chris Bono
  * @since 1.5.0
  */
 @ConfigurationProperties(prefix = "spring.kafka")
-public class KafkaProperties {
-
-	/**
-	 * Comma-delimited list of host:port pairs to use for establishing the initial
-	 * connections to the Kafka cluster. Applies to all components unless overridden.
-	 */
-	private List<String> bootstrapServers = new ArrayList<>(Collections.singletonList("localhost:9092"));
-
-	/**
-	 * ID to pass to the server when making requests. Used for server-side logging.
-	 */
-	private String clientId;
-
-	/**
-	 * Additional properties, common to producers and consumers, used to configure the
-	 * client.
-	 */
-	private final Map<String, String> properties = new HashMap<>();
-
-	private final Consumer consumer = new Consumer();
-
-	private final Producer producer = new Producer();
+public class KafkaProperties extends KafkaPropertiesBaseWithBootstrapServersAndProducerConsumer {
 
 	private final Admin admin = new Admin();
 
@@ -87,42 +54,14 @@ public class KafkaProperties {
 
 	private final Listener listener = new Listener();
 
-	private final Ssl ssl = new Ssl();
-
 	private final Jaas jaas = new Jaas();
 
 	private final Template template = new Template();
 
-	private final Security security = new Security();
-
 	private final Retry retry = new Retry();
 
-	public List<String> getBootstrapServers() {
-		return this.bootstrapServers;
-	}
-
-	public void setBootstrapServers(List<String> bootstrapServers) {
-		this.bootstrapServers = bootstrapServers;
-	}
-
-	public String getClientId() {
-		return this.clientId;
-	}
-
-	public void setClientId(String clientId) {
-		this.clientId = clientId;
-	}
-
-	public Map<String, String> getProperties() {
-		return this.properties;
-	}
-
-	public Consumer getConsumer() {
-		return this.consumer;
-	}
-
-	public Producer getProducer() {
-		return this.producer;
+	public KafkaProperties() {
+		super(new ArrayList<>(Collections.singletonList("localhost:9092")));
 	}
 
 	public Listener getListener() {
@@ -137,10 +76,6 @@ public class KafkaProperties {
 		return this.streams;
 	}
 
-	public Ssl getSsl() {
-		return this.ssl;
-	}
-
 	public Jaas getJaas() {
 		return this.jaas;
 	}
@@ -149,56 +84,8 @@ public class KafkaProperties {
 		return this.template;
 	}
 
-	public Security getSecurity() {
-		return this.security;
-	}
-
 	public Retry getRetry() {
 		return this.retry;
-	}
-
-	private Map<String, Object> buildCommonProperties() {
-		Map<String, Object> properties = new HashMap<>();
-		if (this.bootstrapServers != null) {
-			properties.put(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG, this.bootstrapServers);
-		}
-		if (this.clientId != null) {
-			properties.put(CommonClientConfigs.CLIENT_ID_CONFIG, this.clientId);
-		}
-		properties.putAll(this.ssl.buildProperties());
-		properties.putAll(this.security.buildProperties());
-		if (!CollectionUtils.isEmpty(this.properties)) {
-			properties.putAll(this.properties);
-		}
-		return properties;
-	}
-
-	/**
-	 * Create an initial map of consumer properties from the state of this instance.
-	 * <p>
-	 * This allows you to add additional properties, if necessary, and override the
-	 * default kafkaConsumerFactory bean.
-	 * @return the consumer properties initialized with the customizations defined on this
-	 * instance
-	 */
-	public Map<String, Object> buildConsumerProperties() {
-		Map<String, Object> properties = buildCommonProperties();
-		properties.putAll(this.consumer.buildProperties());
-		return properties;
-	}
-
-	/**
-	 * Create an initial map of producer properties from the state of this instance.
-	 * <p>
-	 * This allows you to add additional properties, if necessary, and override the
-	 * default kafkaProducerFactory bean.
-	 * @return the producer properties initialized with the customizations defined on this
-	 * instance
-	 */
-	public Map<String, Object> buildProducerProperties() {
-		Map<String, Object> properties = buildCommonProperties();
-		properties.putAll(this.producer.buildProperties());
-		return properties;
 	}
 
 	/**
@@ -210,7 +97,11 @@ public class KafkaProperties {
 	 * instance
 	 */
 	public Map<String, Object> buildAdminProperties() {
-		Map<String, Object> properties = buildCommonProperties();
+		// spring.kafka.<common-props>
+		Map<String, Object> properties = buildProperties();
+
+		// spring.kafka.admin.<common-props>
+		// spring.kafka.admin.<specific-props>
 		properties.putAll(this.admin.buildProperties());
 		return properties;
 	}
@@ -223,446 +114,24 @@ public class KafkaProperties {
 	 * instance
 	 */
 	public Map<String, Object> buildStreamsProperties() {
-		Map<String, Object> properties = buildCommonProperties();
+		// spring.kafka.<common-props>
+		Map<String, Object> properties = buildProperties();
+
+		// spring.kafka.streams.<common-props>
+		// spring.kafka.streams.<specific-props>
 		properties.putAll(this.streams.buildProperties());
 		return properties;
 	}
 
-	public static class Consumer {
-
-		private final Ssl ssl = new Ssl();
-
-		private final Security security = new Security();
-
-		/**
-		 * Frequency with which the consumer offsets are auto-committed to Kafka if
-		 * 'enable.auto.commit' is set to true.
-		 */
-		private Duration autoCommitInterval;
-
-		/**
-		 * What to do when there is no initial offset in Kafka or if the current offset no
-		 * longer exists on the server.
-		 */
-		private String autoOffsetReset;
-
-		/**
-		 * Comma-delimited list of host:port pairs to use for establishing the initial
-		 * connections to the Kafka cluster. Overrides the global property, for consumers.
-		 */
-		private List<String> bootstrapServers;
-
-		/**
-		 * ID to pass to the server when making requests. Used for server-side logging.
-		 */
-		private String clientId;
-
-		/**
-		 * Whether the consumer's offset is periodically committed in the background.
-		 */
-		private Boolean enableAutoCommit;
-
-		/**
-		 * Maximum amount of time the server blocks before answering the fetch request if
-		 * there isn't sufficient data to immediately satisfy the requirement given by
-		 * "fetch-min-size".
-		 */
-		private Duration fetchMaxWait;
-
-		/**
-		 * Minimum amount of data the server should return for a fetch request.
-		 */
-		private DataSize fetchMinSize;
-
-		/**
-		 * Unique string that identifies the consumer group to which this consumer
-		 * belongs.
-		 */
-		private String groupId;
-
-		/**
-		 * Expected time between heartbeats to the consumer coordinator.
-		 */
-		private Duration heartbeatInterval;
-
-		/**
-		 * Isolation level for reading messages that have been written transactionally.
-		 */
-		private IsolationLevel isolationLevel = IsolationLevel.READ_UNCOMMITTED;
-
-		/**
-		 * Deserializer class for keys.
-		 */
-		private Class<?> keyDeserializer = StringDeserializer.class;
-
-		/**
-		 * Deserializer class for values.
-		 */
-		private Class<?> valueDeserializer = StringDeserializer.class;
-
-		/**
-		 * Maximum number of records returned in a single call to poll().
-		 */
-		private Integer maxPollRecords;
-
-		/**
-		 * Additional consumer-specific properties used to configure the client.
-		 */
-		private final Map<String, String> properties = new HashMap<>();
-
-		public Ssl getSsl() {
-			return this.ssl;
-		}
-
-		public Security getSecurity() {
-			return this.security;
-		}
-
-		public Duration getAutoCommitInterval() {
-			return this.autoCommitInterval;
-		}
-
-		public void setAutoCommitInterval(Duration autoCommitInterval) {
-			this.autoCommitInterval = autoCommitInterval;
-		}
-
-		public String getAutoOffsetReset() {
-			return this.autoOffsetReset;
-		}
-
-		public void setAutoOffsetReset(String autoOffsetReset) {
-			this.autoOffsetReset = autoOffsetReset;
-		}
-
-		public List<String> getBootstrapServers() {
-			return this.bootstrapServers;
-		}
-
-		public void setBootstrapServers(List<String> bootstrapServers) {
-			this.bootstrapServers = bootstrapServers;
-		}
-
-		public String getClientId() {
-			return this.clientId;
-		}
-
-		public void setClientId(String clientId) {
-			this.clientId = clientId;
-		}
-
-		public Boolean getEnableAutoCommit() {
-			return this.enableAutoCommit;
-		}
-
-		public void setEnableAutoCommit(Boolean enableAutoCommit) {
-			this.enableAutoCommit = enableAutoCommit;
-		}
-
-		public Duration getFetchMaxWait() {
-			return this.fetchMaxWait;
-		}
-
-		public void setFetchMaxWait(Duration fetchMaxWait) {
-			this.fetchMaxWait = fetchMaxWait;
-		}
-
-		public DataSize getFetchMinSize() {
-			return this.fetchMinSize;
-		}
-
-		public void setFetchMinSize(DataSize fetchMinSize) {
-			this.fetchMinSize = fetchMinSize;
-		}
-
-		public String getGroupId() {
-			return this.groupId;
-		}
-
-		public void setGroupId(String groupId) {
-			this.groupId = groupId;
-		}
-
-		public Duration getHeartbeatInterval() {
-			return this.heartbeatInterval;
-		}
-
-		public void setHeartbeatInterval(Duration heartbeatInterval) {
-			this.heartbeatInterval = heartbeatInterval;
-		}
-
-		public IsolationLevel getIsolationLevel() {
-			return this.isolationLevel;
-		}
-
-		public void setIsolationLevel(IsolationLevel isolationLevel) {
-			this.isolationLevel = isolationLevel;
-		}
-
-		public Class<?> getKeyDeserializer() {
-			return this.keyDeserializer;
-		}
-
-		public void setKeyDeserializer(Class<?> keyDeserializer) {
-			this.keyDeserializer = keyDeserializer;
-		}
-
-		public Class<?> getValueDeserializer() {
-			return this.valueDeserializer;
-		}
-
-		public void setValueDeserializer(Class<?> valueDeserializer) {
-			this.valueDeserializer = valueDeserializer;
-		}
-
-		public Integer getMaxPollRecords() {
-			return this.maxPollRecords;
-		}
-
-		public void setMaxPollRecords(Integer maxPollRecords) {
-			this.maxPollRecords = maxPollRecords;
-		}
-
-		public Map<String, String> getProperties() {
-			return this.properties;
-		}
-
-		public Map<String, Object> buildProperties() {
-			Properties properties = new Properties();
-			PropertyMapper map = PropertyMapper.get().alwaysApplyingWhenNonNull();
-			map.from(this::getAutoCommitInterval).asInt(Duration::toMillis)
-					.to(properties.in(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG));
-			map.from(this::getAutoOffsetReset).to(properties.in(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG));
-			map.from(this::getBootstrapServers).to(properties.in(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG));
-			map.from(this::getClientId).to(properties.in(ConsumerConfig.CLIENT_ID_CONFIG));
-			map.from(this::getEnableAutoCommit).to(properties.in(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG));
-			map.from(this::getFetchMaxWait).asInt(Duration::toMillis)
-					.to(properties.in(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG));
-			map.from(this::getFetchMinSize).asInt(DataSize::toBytes)
-					.to(properties.in(ConsumerConfig.FETCH_MIN_BYTES_CONFIG));
-			map.from(this::getGroupId).to(properties.in(ConsumerConfig.GROUP_ID_CONFIG));
-			map.from(this::getHeartbeatInterval).asInt(Duration::toMillis)
-					.to(properties.in(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG));
-			map.from(() -> getIsolationLevel().name().toLowerCase(Locale.ROOT))
-					.to(properties.in(ConsumerConfig.ISOLATION_LEVEL_CONFIG));
-			map.from(this::getKeyDeserializer).to(properties.in(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG));
-			map.from(this::getValueDeserializer).to(properties.in(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG));
-			map.from(this::getMaxPollRecords).to(properties.in(ConsumerConfig.MAX_POLL_RECORDS_CONFIG));
-			return properties.with(this.ssl, this.security, this.properties);
-		}
-
-	}
-
-	public static class Producer {
-
-		private final Ssl ssl = new Ssl();
-
-		private final Security security = new Security();
-
-		/**
-		 * Number of acknowledgments the producer requires the leader to have received
-		 * before considering a request complete.
-		 */
-		private String acks;
-
-		/**
-		 * Default batch size. A small batch size will make batching less common and may
-		 * reduce throughput (a batch size of zero disables batching entirely).
-		 */
-		private DataSize batchSize;
-
-		/**
-		 * Comma-delimited list of host:port pairs to use for establishing the initial
-		 * connections to the Kafka cluster. Overrides the global property, for producers.
-		 */
-		private List<String> bootstrapServers;
-
-		/**
-		 * Total memory size the producer can use to buffer records waiting to be sent to
-		 * the server.
-		 */
-		private DataSize bufferMemory;
-
-		/**
-		 * ID to pass to the server when making requests. Used for server-side logging.
-		 */
-		private String clientId;
-
-		/**
-		 * Compression type for all data generated by the producer.
-		 */
-		private String compressionType;
-
-		/**
-		 * Serializer class for keys.
-		 */
-		private Class<?> keySerializer = StringSerializer.class;
-
-		/**
-		 * Serializer class for values.
-		 */
-		private Class<?> valueSerializer = StringSerializer.class;
-
-		/**
-		 * When greater than zero, enables retrying of failed sends.
-		 */
-		private Integer retries;
-
-		/**
-		 * When non empty, enables transaction support for producer.
-		 */
-		private String transactionIdPrefix;
-
-		/**
-		 * Additional producer-specific properties used to configure the client.
-		 */
-		private final Map<String, String> properties = new HashMap<>();
-
-		public Ssl getSsl() {
-			return this.ssl;
-		}
-
-		public Security getSecurity() {
-			return this.security;
-		}
-
-		public String getAcks() {
-			return this.acks;
-		}
-
-		public void setAcks(String acks) {
-			this.acks = acks;
-		}
-
-		public DataSize getBatchSize() {
-			return this.batchSize;
-		}
-
-		public void setBatchSize(DataSize batchSize) {
-			this.batchSize = batchSize;
-		}
-
-		public List<String> getBootstrapServers() {
-			return this.bootstrapServers;
-		}
-
-		public void setBootstrapServers(List<String> bootstrapServers) {
-			this.bootstrapServers = bootstrapServers;
-		}
-
-		public DataSize getBufferMemory() {
-			return this.bufferMemory;
-		}
-
-		public void setBufferMemory(DataSize bufferMemory) {
-			this.bufferMemory = bufferMemory;
-		}
-
-		public String getClientId() {
-			return this.clientId;
-		}
-
-		public void setClientId(String clientId) {
-			this.clientId = clientId;
-		}
-
-		public String getCompressionType() {
-			return this.compressionType;
-		}
-
-		public void setCompressionType(String compressionType) {
-			this.compressionType = compressionType;
-		}
-
-		public Class<?> getKeySerializer() {
-			return this.keySerializer;
-		}
-
-		public void setKeySerializer(Class<?> keySerializer) {
-			this.keySerializer = keySerializer;
-		}
-
-		public Class<?> getValueSerializer() {
-			return this.valueSerializer;
-		}
-
-		public void setValueSerializer(Class<?> valueSerializer) {
-			this.valueSerializer = valueSerializer;
-		}
-
-		public Integer getRetries() {
-			return this.retries;
-		}
-
-		public void setRetries(Integer retries) {
-			this.retries = retries;
-		}
-
-		public String getTransactionIdPrefix() {
-			return this.transactionIdPrefix;
-		}
-
-		public void setTransactionIdPrefix(String transactionIdPrefix) {
-			this.transactionIdPrefix = transactionIdPrefix;
-		}
-
-		public Map<String, String> getProperties() {
-			return this.properties;
-		}
-
-		public Map<String, Object> buildProperties() {
-			Properties properties = new Properties();
-			PropertyMapper map = PropertyMapper.get().alwaysApplyingWhenNonNull();
-			map.from(this::getAcks).to(properties.in(ProducerConfig.ACKS_CONFIG));
-			map.from(this::getBatchSize).asInt(DataSize::toBytes).to(properties.in(ProducerConfig.BATCH_SIZE_CONFIG));
-			map.from(this::getBootstrapServers).to(properties.in(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG));
-			map.from(this::getBufferMemory).as(DataSize::toBytes)
-					.to(properties.in(ProducerConfig.BUFFER_MEMORY_CONFIG));
-			map.from(this::getClientId).to(properties.in(ProducerConfig.CLIENT_ID_CONFIG));
-			map.from(this::getCompressionType).to(properties.in(ProducerConfig.COMPRESSION_TYPE_CONFIG));
-			map.from(this::getKeySerializer).to(properties.in(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG));
-			map.from(this::getRetries).to(properties.in(ProducerConfig.RETRIES_CONFIG));
-			map.from(this::getValueSerializer).to(properties.in(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG));
-			return properties.with(this.ssl, this.security, this.properties);
-		}
-
-	}
-
-	public static class Admin {
-
-		private final Ssl ssl = new Ssl();
-
-		private final Security security = new Security();
-
-		/**
-		 * ID to pass to the server when making requests. Used for server-side logging.
-		 */
-		private String clientId;
-
-		/**
-		 * Additional admin-specific properties used to configure the client.
-		 */
-		private final Map<String, String> properties = new HashMap<>();
+	/**
+	 * Spring Kafka configuration properties for Kafka admin client.
+	 */
+	public static class Admin extends KafkaPropertiesBase {
 
 		/**
 		 * Whether to fail fast if the broker is not available on startup.
 		 */
 		private boolean failFast;
-
-		public Ssl getSsl() {
-			return this.ssl;
-		}
-
-		public Security getSecurity() {
-			return this.security;
-		}
-
-		public String getClientId() {
-			return this.clientId;
-		}
-
-		public void setClientId(String clientId) {
-			this.clientId = clientId;
-		}
 
 		public boolean isFailFast() {
 			return this.failFast;
@@ -672,55 +141,34 @@ public class KafkaProperties {
 			this.failFast = failFast;
 		}
 
-		public Map<String, String> getProperties() {
-			return this.properties;
-		}
-
 		public Map<String, Object> buildProperties() {
-			Properties properties = new Properties();
-			PropertyMapper map = PropertyMapper.get().alwaysApplyingWhenNonNull();
-			map.from(this::getClientId).to(properties.in(ProducerConfig.CLIENT_ID_CONFIG));
-			return properties.with(this.ssl, this.security, this.properties);
+			// spring.kafka.admin.<common-props>
+			return super.buildProperties();
 		}
 
 	}
 
 	/**
-	 * High (and some medium) priority Streams properties and a general properties bucket.
+	 * Spring Kafka configuration properties for Kafka streams.
 	 */
-	public static class Streams {
-
-		private final Ssl ssl = new Ssl();
-
-		private final Security security = new Security();
+	public static class Streams extends KafkaPropertiesBaseWithBootstrapServers {
 
 		private final Cleanup cleanup = new Cleanup();
 
 		/**
-		 * Kafka streams application.id property; default spring.application.name.
+		 * Kafka streams 'application.id' property; default spring.application.name.
 		 */
 		private String applicationId;
 
 		/**
-		 * Whether or not to auto-start the streams factory bean.
+		 * Whether to auto-start the streams factory bean.
 		 */
 		private boolean autoStartup = true;
-
-		/**
-		 * Comma-delimited list of host:port pairs to use for establishing the initial
-		 * connections to the Kafka cluster. Overrides the global property, for streams.
-		 */
-		private List<String> bootstrapServers;
 
 		/**
 		 * Maximum memory size to be used for buffering across all threads.
 		 */
 		private DataSize cacheMaxSizeBuffering;
-
-		/**
-		 * ID to pass to the server when making requests. Used for server-side logging.
-		 */
-		private String clientId;
 
 		/**
 		 * The replication factor for change log topics and repartition topics created by
@@ -732,19 +180,6 @@ public class KafkaProperties {
 		 * Directory location for the state store.
 		 */
 		private String stateDir;
-
-		/**
-		 * Additional Kafka properties used to configure the streams.
-		 */
-		private final Map<String, String> properties = new HashMap<>();
-
-		public Ssl getSsl() {
-			return this.ssl;
-		}
-
-		public Security getSecurity() {
-			return this.security;
-		}
 
 		public Cleanup getCleanup() {
 			return this.cleanup;
@@ -766,28 +201,12 @@ public class KafkaProperties {
 			this.autoStartup = autoStartup;
 		}
 
-		public List<String> getBootstrapServers() {
-			return this.bootstrapServers;
-		}
-
-		public void setBootstrapServers(List<String> bootstrapServers) {
-			this.bootstrapServers = bootstrapServers;
-		}
-
 		public DataSize getCacheMaxSizeBuffering() {
 			return this.cacheMaxSizeBuffering;
 		}
 
 		public void setCacheMaxSizeBuffering(DataSize cacheMaxSizeBuffering) {
 			this.cacheMaxSizeBuffering = cacheMaxSizeBuffering;
-		}
-
-		public String getClientId() {
-			return this.clientId;
-		}
-
-		public void setClientId(String clientId) {
-			this.clientId = clientId;
 		}
 
 		public Integer getReplicationFactor() {
@@ -806,21 +225,19 @@ public class KafkaProperties {
 			this.stateDir = stateDir;
 		}
 
-		public Map<String, String> getProperties() {
-			return this.properties;
-		}
-
 		public Map<String, Object> buildProperties() {
-			Properties properties = new Properties();
+			// spring.kafka.streams.<common-props>
+			Map<String, Object> properties = super.buildProperties();
+
+			// spring.kafka.streams.<specific-props>
+			PropertiesMap props = new PropertiesMap();
 			PropertyMapper map = PropertyMapper.get().alwaysApplyingWhenNonNull();
-			map.from(this::getApplicationId).to(properties.in("application.id"));
-			map.from(this::getBootstrapServers).to(properties.in(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG));
-			map.from(this::getCacheMaxSizeBuffering).asInt(DataSize::toBytes)
-					.to(properties.in("cache.max.bytes.buffering"));
-			map.from(this::getClientId).to(properties.in(CommonClientConfigs.CLIENT_ID_CONFIG));
-			map.from(this::getReplicationFactor).to(properties.in("replication.factor"));
-			map.from(this::getStateDir).to(properties.in("state.dir"));
-			return properties.with(this.ssl, this.security, this.properties);
+			map.from(this::getApplicationId).to(props.in("application.id"));
+			map.from(this::getCacheMaxSizeBuffering).asInt(DataSize::toBytes).to(props.in("cache.max.bytes.buffering"));
+			map.from(this::getReplicationFactor).to(props.in("replication.factor"));
+			map.from(this::getStateDir).to(props.in("state.dir"));
+			properties.putAll(props);
+			return properties;
 		}
 
 	}
@@ -883,7 +300,7 @@ public class KafkaProperties {
 		private AckMode ackMode;
 
 		/**
-		 * Prefix for the listener's consumer client.id property.
+		 * Prefix for the listener's consumer 'client.id' property.
 		 */
 		private String clientId;
 
@@ -1093,194 +510,6 @@ public class KafkaProperties {
 
 	}
 
-	public static class Ssl {
-
-		/**
-		 * Password of the private key in either key store key or key store file.
-		 */
-		private String keyPassword;
-
-		/**
-		 * Certificate chain in PEM format with a list of X.509 certificates.
-		 */
-		private String keyStoreCertificateChain;
-
-		/**
-		 * Private key in PEM format with PKCS#8 keys.
-		 */
-		private String keyStoreKey;
-
-		/**
-		 * Location of the key store file.
-		 */
-		private Resource keyStoreLocation;
-
-		/**
-		 * Store password for the key store file.
-		 */
-		private String keyStorePassword;
-
-		/**
-		 * Type of the key store.
-		 */
-		private String keyStoreType;
-
-		/**
-		 * Trusted certificates in PEM format with X.509 certificates.
-		 */
-		private String trustStoreCertificates;
-
-		/**
-		 * Location of the trust store file.
-		 */
-		private Resource trustStoreLocation;
-
-		/**
-		 * Store password for the trust store file.
-		 */
-		private String trustStorePassword;
-
-		/**
-		 * Type of the trust store.
-		 */
-		private String trustStoreType;
-
-		/**
-		 * SSL protocol to use.
-		 */
-		private String protocol;
-
-		public String getKeyPassword() {
-			return this.keyPassword;
-		}
-
-		public void setKeyPassword(String keyPassword) {
-			this.keyPassword = keyPassword;
-		}
-
-		public String getKeyStoreCertificateChain() {
-			return this.keyStoreCertificateChain;
-		}
-
-		public void setKeyStoreCertificateChain(String keyStoreCertificateChain) {
-			this.keyStoreCertificateChain = keyStoreCertificateChain;
-		}
-
-		public String getKeyStoreKey() {
-			return this.keyStoreKey;
-		}
-
-		public void setKeyStoreKey(String keyStoreKey) {
-			this.keyStoreKey = keyStoreKey;
-		}
-
-		public Resource getKeyStoreLocation() {
-			return this.keyStoreLocation;
-		}
-
-		public void setKeyStoreLocation(Resource keyStoreLocation) {
-			this.keyStoreLocation = keyStoreLocation;
-		}
-
-		public String getKeyStorePassword() {
-			return this.keyStorePassword;
-		}
-
-		public void setKeyStorePassword(String keyStorePassword) {
-			this.keyStorePassword = keyStorePassword;
-		}
-
-		public String getKeyStoreType() {
-			return this.keyStoreType;
-		}
-
-		public void setKeyStoreType(String keyStoreType) {
-			this.keyStoreType = keyStoreType;
-		}
-
-		public String getTrustStoreCertificates() {
-			return this.trustStoreCertificates;
-		}
-
-		public void setTrustStoreCertificates(String trustStoreCertificates) {
-			this.trustStoreCertificates = trustStoreCertificates;
-		}
-
-		public Resource getTrustStoreLocation() {
-			return this.trustStoreLocation;
-		}
-
-		public void setTrustStoreLocation(Resource trustStoreLocation) {
-			this.trustStoreLocation = trustStoreLocation;
-		}
-
-		public String getTrustStorePassword() {
-			return this.trustStorePassword;
-		}
-
-		public void setTrustStorePassword(String trustStorePassword) {
-			this.trustStorePassword = trustStorePassword;
-		}
-
-		public String getTrustStoreType() {
-			return this.trustStoreType;
-		}
-
-		public void setTrustStoreType(String trustStoreType) {
-			this.trustStoreType = trustStoreType;
-		}
-
-		public String getProtocol() {
-			return this.protocol;
-		}
-
-		public void setProtocol(String protocol) {
-			this.protocol = protocol;
-		}
-
-		public Map<String, Object> buildProperties() {
-			validate();
-			Properties properties = new Properties();
-			PropertyMapper map = PropertyMapper.get().alwaysApplyingWhenNonNull();
-			map.from(this::getKeyPassword).to(properties.in(SslConfigs.SSL_KEY_PASSWORD_CONFIG));
-			map.from(this::getKeyStoreCertificateChain)
-					.to(properties.in(SslConfigs.SSL_KEYSTORE_CERTIFICATE_CHAIN_CONFIG));
-			map.from(this::getKeyStoreKey).to(properties.in(SslConfigs.SSL_KEYSTORE_KEY_CONFIG));
-			map.from(this::getKeyStoreLocation).as(this::resourceToPath)
-					.to(properties.in(SslConfigs.SSL_KEYSTORE_LOCATION_CONFIG));
-			map.from(this::getKeyStorePassword).to(properties.in(SslConfigs.SSL_KEYSTORE_PASSWORD_CONFIG));
-			map.from(this::getKeyStoreType).to(properties.in(SslConfigs.SSL_KEYSTORE_TYPE_CONFIG));
-			map.from(this::getTrustStoreCertificates).to(properties.in(SslConfigs.SSL_TRUSTSTORE_CERTIFICATES_CONFIG));
-			map.from(this::getTrustStoreLocation).as(this::resourceToPath)
-					.to(properties.in(SslConfigs.SSL_TRUSTSTORE_LOCATION_CONFIG));
-			map.from(this::getTrustStorePassword).to(properties.in(SslConfigs.SSL_TRUSTSTORE_PASSWORD_CONFIG));
-			map.from(this::getTrustStoreType).to(properties.in(SslConfigs.SSL_TRUSTSTORE_TYPE_CONFIG));
-			map.from(this::getProtocol).to(properties.in(SslConfigs.SSL_PROTOCOL_CONFIG));
-			return properties;
-		}
-
-		private void validate() {
-			MutuallyExclusiveConfigurationPropertiesException.throwIfMultipleNonNullValuesIn((entries) -> {
-				entries.put("spring.kafka.ssl.key-store-key", this.getKeyStoreKey());
-				entries.put("spring.kafka.ssl.key-store-location", this.getKeyStoreLocation());
-			});
-			MutuallyExclusiveConfigurationPropertiesException.throwIfMultipleNonNullValuesIn((entries) -> {
-				entries.put("spring.kafka.ssl.trust-store-certificates", this.getTrustStoreCertificates());
-				entries.put("spring.kafka.ssl.trust-store-location", this.getTrustStoreLocation());
-			});
-		}
-
-		private String resourceToPath(Resource resource) {
-			try {
-				return resource.getFile().getAbsolutePath();
-			}
-			catch (IOException ex) {
-				throw new IllegalStateException("Resource '" + resource + "' must be on a file system", ex);
-			}
-		}
-
-	}
-
 	public static class Jaas {
 
 		/**
@@ -1335,30 +564,6 @@ public class KafkaProperties {
 			if (options != null) {
 				this.options.putAll(options);
 			}
-		}
-
-	}
-
-	public static class Security {
-
-		/**
-		 * Security protocol used to communicate with brokers.
-		 */
-		private String protocol;
-
-		public String getProtocol() {
-			return this.protocol;
-		}
-
-		public void setProtocol(String protocol) {
-			this.protocol = protocol;
-		}
-
-		public Map<String, Object> buildProperties() {
-			Properties properties = new Properties();
-			PropertyMapper map = PropertyMapper.get().alwaysApplyingWhenNonNull();
-			map.from(this::getProtocol).to(properties.in(CommonClientConfigs.SECURITY_PROTOCOL_CONFIG));
-			return properties;
 		}
 
 	}
@@ -1487,47 +692,6 @@ public class KafkaProperties {
 
 		public void setOnShutdown(boolean onShutdown) {
 			this.onShutdown = onShutdown;
-		}
-
-	}
-
-	public enum IsolationLevel {
-
-		/**
-		 * Read everything including aborted transactions.
-		 */
-		READ_UNCOMMITTED((byte) 0),
-
-		/**
-		 * Read records from committed transactions, in addition to records not part of
-		 * transactions.
-		 */
-		READ_COMMITTED((byte) 1);
-
-		private final byte id;
-
-		IsolationLevel(byte id) {
-			this.id = id;
-		}
-
-		public byte id() {
-			return this.id;
-		}
-
-	}
-
-	@SuppressWarnings("serial")
-	private static class Properties extends HashMap<String, Object> {
-
-		<V> java.util.function.Consumer<V> in(String key) {
-			return (value) -> put(key, value);
-		}
-
-		Properties with(Ssl ssl, Security security, Map<String, String> properties) {
-			putAll(ssl.buildProperties());
-			putAll(security.buildProperties());
-			putAll(properties);
-			return this;
 		}
 
 	}
