+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/kafka/PropertiesMap.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2022-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.boot.autoconfigure.kafka;
+
+import java.util.HashMap;
+
+/**
+ * Map of properties that provides a {@link #in convenience method} to consume a property entry.
+ *
+ * @author Chris Bono
+ */
+class PropertiesMap extends HashMap<String, Object> {
+
+	<V> java.util.function.Consumer<V> in(String key) {
+		return (value) -> put(key, value);
+	}
+
+}
